Split fuzzer.rs into smaller files, rename `RandomFuzzer` to `Random`
