+++ crates/forge/src/fuzzer.rs
@@ -1,118 +1,12 @@
-use anyhow::anyhow;
+use crate::fuzzer::arguments::CairoType;
 use anyhow::Result;
-use cairo_felt::Felt252;
-use num_bigint::{BigUint, RandBigInt};
-use num_integer::Integer;
-use num_traits::{One, Zero};
 use rand::rngs::StdRng;
-use rand::{Rng, SeedableRng};
-use std::ops::{Add, Shl, Shr, Sub};
+use rand::Rng;
 
-#[derive(Debug, Clone)]
-enum CairoType {
-    U8,
-    U16,
-    U32,
-    U64,
-    U128,
-    U256,
-    Felt252,
-}
+mod arguments;
+mod random;
 
-trait Argument {
-    fn low(&self) -> BigUint;
-    fn high(&self) -> BigUint;
-    fn gen(&self, rng: &mut StdRng) -> Vec<Felt252>;
-    fn min(&self) -> Vec<Felt252>;
-    fn max(&self) -> Vec<Felt252>;
-}
-
-impl Argument for CairoType {
-    fn low(&self) -> BigUint {
-        BigUint::zero()
-    }
-
-    fn high(&self) -> BigUint {
-        match self {
-            CairoType::U8 => BigUint::from(u8::MAX).add(BigUint::one()),
-            CairoType::U16 => BigUint::from(u16::MAX).add(BigUint::one()),
-            CairoType::U32 => BigUint::from(u32::MAX).add(BigUint::one()),
-            CairoType::U64 => BigUint::from(u64::MAX).add(BigUint::one()),
-            CairoType::U128 => BigUint::from(u128::MAX).add(BigUint::one()),
-            CairoType::U256 => {
-                let max = BigUint::from(1_u32);
-                max.shl(256)
-            }
-            CairoType::Felt252 => Felt252::prime(),
-        }
-    }
-
-    fn gen(&self, rng: &mut StdRng) -> Vec<Felt252> {
-        match self {
-            CairoType::U8
-            | CairoType::U16
-            | CairoType::U32
-            | CairoType::U64
-            | CairoType::U128
-            | CairoType::Felt252 => {
-                vec![Felt252::from(
-                    rng.gen_biguint_range(&self.low(), &self.high()),
-                )]
-            }
-            CairoType::U256 => {
-                let val = rng.gen_biguint_range(&self.low(), &self.high());
-                u256_to_felt252(val)
-            }
-        }
-    }
-
-    fn min(&self) -> Vec<Felt252> {
-        match self {
-            CairoType::U8
-            | CairoType::U16
-            | CairoType::U32
-            | CairoType::U64
-            | CairoType::U128
-            | CairoType::Felt252 => vec![Felt252::from(self.low())],
-            CairoType::U256 => vec![Felt252::from(self.low()), Felt252::from(self.low())],
-        }
-    }
-
-    fn max(&self) -> Vec<Felt252> {
-        match self {
-            CairoType::U8
-            | CairoType::U16
-            | CairoType::U32
-            | CairoType::U64
-            | CairoType::U128
-            | CairoType::Felt252 => vec![Felt252::from(self.high().sub(BigUint::one()))],
-            CairoType::U256 => u256_to_felt252(self.high().sub(BigUint::one())),
-        }
-    }
-}
-
-fn u256_to_felt252(val: BigUint) -> Vec<Felt252> {
-    let low = val.mod_floor(&BigUint::from(2_u32).pow(128));
-    let high = val.shr(128);
-    vec![Felt252::from(low), Felt252::from(high)]
-}
-
-impl CairoType {
-    fn from_name(name: &str) -> Result<Self> {
-        match name {
-            "u8" => Ok(Self::U8),
-            "u16" => Ok(Self::U16),
-            "u32" => Ok(Self::U32),
-            "u64" => Ok(Self::U64),
-            "u128" => Ok(Self::U128),
-            "u256" | "core::integer::u256" => Ok(Self::U256),
-            "felt252" => Ok(Self::Felt252),
-            _ => Err(anyhow!(
-                "Tried to use incorrect type for fuzzing. Type = {name} is not supported"
-            )),
-        }
-    }
-}
+pub use random::Random;
 
 #[derive(Debug, Clone)]
 pub struct RunParams {
@@ -163,193 +57,5 @@ impl RunParams {
             run_with_min_value_for_argument,
             run_with_max_value_for_argument,
         })
-    }
-}
-
-#[derive(Debug, Clone)]
-#[allow(clippy::module_name_repetitions)]
-pub struct RandomFuzzer {
-    rng: StdRng,
-    run_params: RunParams,
-}
-
-impl RandomFuzzer {
-    pub fn create(seed: u64, total_runs: u32, arguments: &[&str]) -> Result<Self> {
-        let mut rng = StdRng::seed_from_u64(seed);
-        let run_params = RunParams::from(&mut rng, total_runs, arguments)?;
-
-        Ok(RandomFuzzer { rng, run_params })
-    }
-
-    pub fn next_args(&mut self) -> Vec<Felt252> {
-        assert!(self.run_params.executed_runs < self.run_params.total_runs);
-
-        self.next_run();
-
-        let mut args = vec![];
-
-        for (index, argument) in self.run_params.arguments.iter().enumerate() {
-            if self.is_run_with_min_value_for_arg(index) {
-                args.extend(argument.min());
-            } else if self.is_run_with_max_value_for_arg(index) {
-                args.extend(argument.max());
-            } else {
-                args.extend(argument.gen(&mut self.rng));
-            }
-        }
-
-        args
-    }
-
-    fn next_run(&mut self) {
-        self.run_params.executed_runs += 1;
-    }
-
-    fn is_run_with_min_value_for_arg(&self, arg_number: usize) -> bool {
-        let current_run = self.run_params.executed_runs;
-        self.run_params.run_with_min_value_for_argument[arg_number] == current_run
-    }
-
-    fn is_run_with_max_value_for_arg(&self, arg_number: usize) -> bool {
-        let current_run = self.run_params.executed_runs;
-        self.run_params.run_with_max_value_for_argument[arg_number] == current_run
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rand::{thread_rng, RngCore};
-
-    impl Default for RunParams {
-        fn default() -> Self {
-            Self {
-                arguments: vec![],
-                total_runs: 256,
-                executed_runs: 0,
-                run_with_min_value_for_argument: vec![],
-                run_with_max_value_for_argument: vec![],
-            }
-        }
-    }
-
-    fn all_values_different(vec1: &[Felt252], vec2: &[Felt252]) -> bool {
-        vec1.iter().zip(vec2).all(|(i, j)| {
-            i.to_biguint() != j.to_biguint()
-                && i.to_biguint() != BigUint::zero()
-                && j.to_biguint() != BigUint::zero()
-        })
-    }
-
-    #[test]
-    fn fuzzer_generates_different_values() {
-        let run_params = RunParams {
-            run_with_max_value_for_argument: vec![3, 3, 3],
-            run_with_min_value_for_argument: vec![4, 4, 4],
-            arguments: vec![CairoType::Felt252, CairoType::U256, CairoType::U32],
-            total_runs: 10,
-            ..Default::default()
-        };
-        let mut fuzzer = RandomFuzzer {
-            rng: StdRng::seed_from_u64(1234),
-            run_params,
-        };
-
-        let args1 = fuzzer.next_args();
-        let args2 = fuzzer.next_args();
-        let args3 = fuzzer.next_args();
-
-        assert!(all_values_different(&args1, &args2));
-        assert!(all_values_different(&args1, &args3));
-        assert!(all_values_different(&args2, &args3));
-    }
-
-    #[test]
-    fn run_with_max_value() {
-        let run_params = RunParams {
-            run_with_max_value_for_argument: vec![1],
-            run_with_min_value_for_argument: vec![2],
-            total_runs: 10,
-            ..Default::default()
-        };
-        let mut fuzzer = RandomFuzzer {
-            rng: StdRng::seed_from_u64(1234),
-            run_params,
-        };
-
-        assert!(!fuzzer.is_run_with_max_value_for_arg(0));
-        fuzzer.next_args();
-        assert!(fuzzer.is_run_with_max_value_for_arg(0));
-    }
-
-    #[test]
-    fn run_with_min_value() {
-        let run_params = RunParams {
-            run_with_max_value_for_argument: vec![2],
-            run_with_min_value_for_argument: vec![1],
-            total_runs: 10,
-            ..Default::default()
-        };
-        let mut fuzzer = RandomFuzzer {
-            rng: StdRng::seed_from_u64(1234),
-            run_params,
-        };
-
-        assert!(!fuzzer.is_run_with_min_value_for_arg(0));
-        fuzzer.next_args();
-        assert!(fuzzer.is_run_with_min_value_for_arg(0));
-    }
-
-    #[test]
-    fn using_seed_consistent_result() {
-        let seed = thread_rng().next_u64();
-        let mut fuzzer = RandomFuzzer::create(seed, 3, &["felt252", "felt252", "felt252"]).unwrap();
-        let values = fuzzer.next_args();
-
-        let mut fuzzer = RandomFuzzer::create(seed, 3, &["felt252", "felt252", "felt252"]).unwrap();
-        let values_from_seed = fuzzer.next_args();
-
-        assert_eq!(values, values_from_seed);
-    }
-
-    #[test]
-    fn min_and_max_used_at_least_once_for_each_arg() {
-        let seed = thread_rng().next_u64();
-        let runs_number = 10;
-        let arguments = vec!["felt252", "felt252", "felt252"];
-        let args_number = arguments.len();
-
-        let mut fuzzer = RandomFuzzer::create(seed, runs_number, &arguments).unwrap();
-
-        let mut min_used = vec![false; args_number];
-        let mut max_used = vec![false; args_number];
-
-        for _ in 1..=runs_number {
-            let values = fuzzer.next_args();
-            for (i, value) in values.iter().enumerate() {
-                assert!(
-                    *value >= CairoType::Felt252.min()[0] && *value <= CairoType::Felt252.max()[0]
-                );
-                if *value == CairoType::Felt252.min()[0] {
-                    min_used[i] = true;
-                } else if *value == CairoType::Felt252.max()[0] {
-                    max_used[i] = true;
-                }
-            }
-        }
-
-        assert_eq!(min_used, vec![true; args_number]);
-        assert_eq!(max_used, vec![true; args_number]);
-    }
-
-    #[test]
-    fn create_fuzzer_from_invalid_arguments() {
-        let result = RandomFuzzer::create(1234, 512, &["felt252", "invalid", "args"]);
-        let err = result.unwrap_err();
-
-        assert_eq!(
-            err.to_string(),
-            "Tried to use incorrect type for fuzzing. Type = invalid is not supported"
-        );
     }
 }
